+++ cogs/__init__.py
@@ -1,0 +1,12 @@
+"""
+Required for cogs to find parent modules
+
+as of commit 4ab98bb, this may not be needed
+https://github.com/V3ntus/repo-finder-bot/commit/4ab98bbad75c398a3214da36a4079f1548d4b6f1
+"""
+
+# import sys
+# from os.path import dirname as dn
+
+# def __init__(self):
+#     sys.path.insert(0, dn(sys.path[0])))
