+++ slashcogs/template.py
